requirements.txt
# requirements.txt — +blog autonomo
# Dependências extraídas dos imports de todos os scripts em execution/

python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.119.0
